requirements.txt
beautifulsoup4==4.12.3
Flask==2.2.5
nltk==3.8.1
numpy==1.24.3
pandas==2.2.2
Pillow==10.3.0
Requests==2.31.0
scikit_learn==1.3.0
scipy==1.11.4
tqdm==4.65.0
